PaymentWalletJDBC/src/main/java/com/capgemini/bean/CustomerDetails.java
package com.capgemini.bean;

public class CustomerDetails {

	

	private String firstName;
	private String lastName;
	@Override
	public String toString() {
		return "Customer [firstName=" + firstName + ", lastName=" + lastName + ", gender=" + gender + ", age=" + age
				+ ", mobileNo=" + mobileNo + ", email=" + email + ", aadhar=" + aadhar + ", location=" + location + "]";
	}
	public String getFirstName() {
		return firstName;
	}
	public void setFirstName(String firstName) {
		this.firstName = firstName;
	}
	public String getLastName() {
		return lastName;
	}
	public void setLastName(String lastName) {
		this.lastName = lastName;
	}
	public String getGender() {
		return gender;
	}
	public void setGender(String gender) {
		this.gender = gender;
	}
	public int getAge() {
		return age;
	}
	public void setAge(int age) {
		this.age = age;
	}
	
	public Long getMobileNo() {
		return mobileNo;
	}
	public void setMobileNo(Long mobileNo) {
		this.mobileNo = mobileNo;
	}
	public String getEmail() {
		return email;
	}
	public void setEmail(String email) {
		this.email = email;
	}
	
	public String getLocation() {
		return location;
	}
	public void setLocation(String location) {
		this.location = location;
	}
	private String gender;
	private int age;
	public long getAadhar() {
		return aadhar;
	}
	public void setAadhar(long aadhar) {
		this.aadhar = aadhar;
	}
	private Long mobileNo;
	private String email;
	private long aadhar;
	private String location;
	
	
	
	
	
	
	
	
	
	
	
	
	
	
	
	
	
	
	
	
}
